requirements.txt
# requirements.txt

# Core Streamlit
streamlit>=1.28.0

# Interactive components for image editing
streamlit-drawable-canvas>=0.9.3
streamlit-image-coordinates>=0.1.6
plotly>=5.15.0

# ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0

# Data handling
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
pathlib>=1.0.0
streamlit-drawable-canvas
# Optional: For better performance
# streamlit-aggrid>=0.3.4  # For better data tables